资源详情接口访问量统计

+++ src/main/java/cc/mrbird/febs/resource/entity/Resource.java
@@ -134,7 +134,7 @@ public class Resource {
      * 阅读数
      */
     @TableField("read_count")
-    private Integer readCount;
+    private Long readCount;
 
     /**
      * 评论数
